start fleshing out api
