+++ raphtory/src/db/mutation_api/deletion_ops.rs
@@ -1,1 +1,25 @@
-pub trait DeletionOps {}
+use crate::core::tgraph_shard::errors::GraphError;
+use crate::core::time::{IntoTimeWithFormat, TryIntoTime};
+use crate::core::vertex::InputVertex;
+
+pub trait DeletionOps {
+    fn delete_edge<V: InputVertex, T: TryIntoTime>(
+        &self,
+        t: T,
+        src: V,
+        dst: V,
+        layer: Option<&str>,
+    ) -> Result<(), GraphError>;
+
+    fn delete_edge_with_custom_time_format<V: InputVertex>(
+        &self,
+        t: &str,
+        fmt: &str,
+        src: V,
+        dst: V,
+        layer: Option<&str>,
+    ) -> Result<(), GraphError> {
+        let time: i64 = t.parse_time(fmt)?;
+        self.delete_edge(time, src, dst, layer)
+    }
+}
