+++ raphtory/src/db/mutation_api/mod.rs
@@ -1,5 +1,7 @@
 mod addition_ops;
 mod deletion_ops;
+mod property_addition_ops;
 
 pub use addition_ops::AdditionOps;
 pub use deletion_ops::DeletionOps;
+pub use property_addition_ops::PropertyAdditionOps;
